Remove datastore
